refactor: split back_log and save into another file
What's a terrible usage of global variables!
