+++ LINKS.vcxproj
@@ -19,18 +19,22 @@
     </ProjectConfiguration>
   </ItemGroup>
   <ItemGroup>
+    <ClCompile Include="back_log.cpp" />
     <ClCompile Include="fmt\fmt\format.cc" />
     <ClCompile Include="GAME.cpp" />
     <ClCompile Include="LINKSMAIN.cpp" />
     <ClCompile Include="resource_manager.cpp" />
+    <ClCompile Include="save.cpp" />
     <ClCompile Include="utility.cpp" />
   </ItemGroup>
   <ItemGroup>
+    <ClInclude Include="back_log.hpp" />
     <ClInclude Include="DEF.h" />
     <ClInclude Include="fmt\fmt\format.h" />
     <ClInclude Include="GAME.h" />
     <ClInclude Include="resource.h" />
     <ClInclude Include="resource_manager.hpp" />
+    <ClInclude Include="save.hpp" />
     <ClInclude Include="utility.hpp" />
   </ItemGroup>
   <ItemGroup>
